+++ src/pipelines/mod.rs
@@ -44,8 +44,8 @@
 //! - English <-> Catalan
 //! - English <-> German
 //! - English <-> Russian
-//! - English <-> Mandarin
-//! - English <-> Cantonese
+//! - English <-> Chinese (Simplified)
+//! - English <-> Chinese (Traditional)
 //! - English <-> Dutch
 //! - English <-> Swedish
 //! - English <-> Arabic